requirements.txt
pymilvus 
sentence-transformers
python-dotenv
google-genai